requirements.txt
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-PyMongo==2.3.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymongo==4.10.1
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.3
uWSGI==2.0.28
Werkzeug==3.1.3
